moving securityUtils to core
